set1

+++ set1/6/6.c
@@ -145,7 +145,7 @@ main(void)
 	FILE *memstream;
 	char *buf, tmp[BUFSIZ], *key;
 	size_t i, len, keylen;
-	int nr;
+	ssize_t nr;
 
 	if ((bio = BIO_new_fp(stdin, BIO_NOCLOSE)) == NULL ||
 	    (b64 = BIO_new(BIO_f_base64())) == NULL ||
