+++ set1/7/7.c
@@ -11,7 +11,7 @@ main(void)
 {
 	BIO *bio, *b64, *cip;
 	char buf[BUFSIZ];
-	int nr;
+	ssize_t nr;
 
 	if ((bio = BIO_new_fp(stdin, BIO_NOCLOSE)) == NULL ||
 	    (b64 = BIO_new(BIO_f_base64())) == NULL ||
